feat(pano_image_publisher): Allow for less than 1hz publish rate

+++ main_ws/src/pano_image_publisher/include/pano_image_publisher/publisher_node.hpp
@@ -28,7 +28,7 @@ private:
   // Params
   std::string folder_path_;
   std::string publish_topic_name_;
-  int publish_rate_in_hz_;
+  float publish_rate_in_hz_;
   bool loop_;
   bool shuffle_;
